+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes for Craft Commerce
 
-## Unreleased
+## 5.0.3 - 2024-05-02
 
 - Added `craft\commerce\helpers\ProjectConfigData::ensureAllStoresProcessed()`.
 - Added `craft\commerce\models\OrderStatus::getConfig()`.